third_party/WebKit/LayoutTests/fast/css/content-property-serialization.html

<!doctype html>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../css-parser/resources/property-parsing-test.js"></script>
<script>
description('Verifies that a css text serialization of the content property can be roundtripped.');
assert_valid_value("content", "' (' url(http://example.com/) ')'");
</script>
